prevent the deletion of an availability when attending an event (Fixes #45)

+++ webservice/da/src/main/java/net/cpollet/itinerants/da/neo4j/service/Neo4jAvailabilityService.java
@@ -2,8 +2,12 @@ package net.cpollet.itinerants.da.neo4j.service;
 
 import lombok.extern.slf4j.Slf4j;
 import net.cpollet.itinerants.core.service.AvailabilityService;
+import net.cpollet.itinerants.da.neo4j.data.Neo4JPersonData;
+import net.cpollet.itinerants.da.neo4j.repositories.AttendanceRepository;
 import net.cpollet.itinerants.da.neo4j.repositories.AvailabilityRepository;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 /**
  * Created by cpollet on 13.02.17.
@@ -12,9 +16,11 @@ import org.springframework.stereotype.Service;
 @Slf4j
 public class Neo4jAvailabilityService implements AvailabilityService {
     private final AvailabilityRepository availabilityRepository;
+    private final AttendanceRepository attendanceRepository;
 
-    public Neo4jAvailabilityService(AvailabilityRepository availabilityRepository) {
+    public Neo4jAvailabilityService(AvailabilityRepository availabilityRepository, AttendanceRepository attendanceRepository) {
         this.availabilityRepository = availabilityRepository;
+        this.attendanceRepository = attendanceRepository;
     }
 
     @Override
@@ -24,6 +30,13 @@ public class Neo4jAvailabilityService implements AvailabilityService {
 
     @Override
     public void delete(InputAvailabilityData availability) {
-        availabilityRepository.delete(availability.getPersonId(), availability.getEventId());
+        if (!isAttending(availability.getPersonId(), availability.getEventId())) {
+            availabilityRepository.delete(availability.getPersonId(), availability.getEventId());
+        }
+    }
+
+    private boolean isAttending(String personId, String eventId) {
+        List<Neo4JPersonData> attendees = attendanceRepository.attendees(eventId);
+        return attendees.stream().anyMatch(p -> p.getUUID().equals(personId));
     }
 }
